+++ src/main/java/com/supernova/fashionnova/order/dto/OrderDetailResponseDto.java
@@ -10,13 +10,12 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 public class OrderDetailResponseDto {
-
   private Long id;
   private int count;
   private String productName;
-  private int price;
+  private Long price;
 
-  public OrderDetailResponseDto(Long id, int count, String productName, int price) {
+  public OrderDetailResponseDto(Long id, int count, String productName, Long price) {
     this.id = id;
     this.count = count;
     this.productName = productName;
